+++ pkg/convert/python/python_test.go
@@ -68,6 +68,7 @@ func TestGetPythonMeta(t *testing.T) {
 
 	pythonctx, err := New("botocore")
 	pythonctx.PackageIndex = NewPackageIndex("https://pypi.org")
+	pythonctx.PackageIndex.Client.Ratelimiter = nil // don't rate limit our unit tests
 	assert.NoError(t, err)
 
 	var expected Package
@@ -106,6 +107,7 @@ func TestFindDependencies(t *testing.T) {
 
 		for _, pythonctx := range pythonctxs {
 			pythonctx.PackageIndex.url = server.URL
+			pythonctx.PackageIndex.Client.Ratelimiter = nil // don't rate limit our unit tests
 			p, err := pythonctx.PackageIndex.Get(slogtest.TestContextWithLogger(t), pythonctx.PackageName, pythonctx.PackageVersion)
 			assert.NoError(t, err)
 			pythonctx.ToCheck = append(pythonctx.ToCheck, p.Info.Name)
@@ -130,73 +132,6 @@ func TestFindDependencies(t *testing.T) {
 			// The dependency list should be empty
 			assert.Empty(t, pythonctx.ToGenerate)
 		}
-	}
-}
-
-func TestGenerateManifest(t *testing.T) {
-	ctx := slogtest.TestContextWithLogger(t)
-
-	for i := range versions {
-		pythonctxs, err := SetupContext(versions[i])
-		assert.NoError(t, err)
-
-		// botocore ctx
-		pythonctx := pythonctxs[0]
-		// Add additionalReposities and additionalKeyrings
-		pythonctx.AdditionalRepositories = []string{"https://packages.wolfi.dev/os"}
-		pythonctx.AdditionalKeyrings = []string{"https://packages.wolfi.dev/os/wolfi-signing.rsa.pub"}
-
-		got, err := pythonctx.generateManifest(ctx, pythonctx.Package, pythonctx.PackageVersion, nil, nil)
-		assert.NoError(t, err)
-
-		// Check Package
-		assert.Equal(t, got.Package.Name, "py"+versions[i]+"-botocore")
-		assert.Equal(t, got.Package.Version, "1.29.78")
-		assert.EqualValues(t, got.Package.Epoch, 0)
-		assert.Equal(t, got.Package.Description, "Low-level, data-driven core of boto 3.")
-		assert.Equal(t, got.Package.Dependencies.Runtime, []string{"py" + versions[i] + "-jmespath", "py" + versions[i] + "-python-dateutil", "py" + versions[i] + "-urllib3", "python-" + versions[i]})
-
-		// Check Package.Copyright
-		assert.Equal(t, len(got.Package.Copyright), 1)
-		assert.Equal(t, got.Package.Copyright[0].License, "Apache License 2.0")
-
-		// Check Environment
-		assert.Equal(t, got.Environment.Contents.Packages, []string{
-			"build-base",
-			"busybox",
-			"ca-certificates-bundle",
-			"wolfi-base",
-		})
-
-		// Check Pipeline
-		assert.Equal(t, len(got.Pipeline), 3)
-
-		// Check Pipeline - fetch
-		assert.Equal(t, got.Pipeline[0].Uses, "fetch")
-
-		releases, ok := pythonctx.Package.Releases[pythonctx.PackageVersion]
-
-		// If the key exists
-		assert.True(t, ok)
-
-		var release Release
-		for _, r := range releases {
-			if r.PackageType == "sdist" {
-				release = r
-			}
-		}
-		assert.NotEmpty(t, release)
-		assert.Equal(t, "https://files.pythonhosted.org/packages/8f/34/d4bcefeabfb8e4b46157e84ea55c3ecc7399d5f9a3454728e1d0d5f9cb83/botocore-"+pythonctx.PackageVersion+".tar.gz", release.URL)
-
-		tempURI := fmt.Sprintf("https://files.pythonhosted.org/packages/source/%c/%s/%s-%s.tar.gz", pythonctx.PackageName[0], pythonctx.PackageName, pythonctx.PackageName, pythonctx.PackageVersion)
-		assert.Equal(t, got.Pipeline[0].With, map[string]string{
-			"expected-sha256": "2bee6ed037590ef1e4884d944486232871513915f12a8590c63e3bb6046479bf",
-			"uri":             strings.ReplaceAll(tempURI, pythonctx.PackageVersion, "${{package.version}}"),
-		})
-
-		// Check Pipeline - runs
-		assert.Equal(t, got.Pipeline[1].Uses, "python/build-wheel")
-		assert.Equal(t, got.Pipeline[2].Uses, "strip")
 	}
 }
 
@@ -236,6 +171,7 @@ func SetupContext(version string) ([]*PythonContext, error) {
 	}
 
 	botocorepythonctx.PackageIndex = NewPackageIndex("https://pypi.org/")
+	botocorepythonctx.PackageIndex.Client.Ratelimiter = nil // don't rate limit our unit tests
 	botocorepythonctx.PackageName = "botocore"
 	botocorepythonctx.PackageVersion = "1.29.78"
 	botocorepythonctx.PythonVersion = version
@@ -261,6 +197,7 @@ func SetupContext(version string) ([]*PythonContext, error) {
 	}
 
 	jsonschemapythonctx.PackageIndex = NewPackageIndex("https://pypi.org/")
+	jsonschemapythonctx.PackageIndex.Client.Ratelimiter = nil // don't rate limit our unit tests
 	jsonschemapythonctx.PackageName = "jsonschema"
 	jsonschemapythonctx.PackageVersion = "4.17.3"
 	jsonschemapythonctx.PythonVersion = version
