+++ pkg/http/http.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"net/http"
 
+	"github.com/chainguard-dev/clog"
 	"golang.org/x/time/rate"
 )
 
@@ -18,10 +19,12 @@ type RLHTTPClient struct {
 
 // Do dispatches the HTTP request to the network
 func (c *RLHTTPClient) Do(req *http.Request) (*http.Response, error) {
-	// Comment out the below 5 lines to turn off ratelimiting
-	err := c.Ratelimiter.Wait(req.Context()) // This is a blocking call. Honors the rate limit
-	if err != nil {
-		return nil, err
+	if c.Ratelimiter != nil {
+		// Comment out the below 5 lines to turn off ratelimiting
+		err := c.Ratelimiter.Wait(req.Context()) // This is a blocking call. Honors the rate limit
+		if err != nil {
+			return nil, err
+		}
 	}
 	resp, err := c.Client.Do(req)
 	if err != nil {
@@ -44,12 +47,14 @@ func NewClient(rl *rate.Limiter) *RLHTTPClient {
 //
 // On success, it will return the sha256 hash as a string.
 func (c *RLHTTPClient) GetArtifactSHA256(ctx context.Context, artifactURI string) (string, error) {
+	clog.FromContext(ctx).Infof("Getting artifact %s", artifactURI)
 	req, err := http.NewRequestWithContext(ctx, "GET", artifactURI, nil)
 	if err != nil {
 		return "", fmt.Errorf("creating request for %s: %w", artifactURI, err)
 	}
 	var client http.Client
 
+	clog.FromContext(ctx).Infof("client.Do(%q)", artifactURI)
 	resp, err := client.Do(req)
 	if err != nil {
 		return "", fmt.Errorf("getting %s: %w", artifactURI, err)
